Rename FilteredVariantsList
